spark_bloomFilter_alternative

+++ spark-bloom-filter/bloomfilters_util.py
@@ -15,7 +15,7 @@ def compute_hashes(line: str, size_of_bloom_filters: dict, hash_function_number:
     """
     # Parse line
     movie_id = line[0]
-    rating = int(round(float(line[1])))
+    rating = round(float(line[1]))
     # Get bloom filter size
     bloom_filter_size = size_of_bloom_filters[rating]
     # Compute hash values
@@ -30,7 +30,7 @@ def compute_number_of_hash_functions(false_positive_prob: float) -> int:
     
     :return: the number of hash functions
     """
-    return int(math.ceil(- math.log(false_positive_prob) / math.log(2)))
+    return math.ceil(- math.log(false_positive_prob) / math.log(2))
 
 
 def compute_size_of_bloom_filter(false_positive_prob: float, number_of_inputs: int) -> int:
@@ -42,7 +42,7 @@ def compute_size_of_bloom_filter(false_positive_prob: float, number_of_inputs: i
     
     :return: the size of a single bloom filter
     """
-    return int(math.ceil(-(number_of_inputs * math.log(false_positive_prob)) / math.pow(math.log(2), 2)))
+    return math.ceil(-(number_of_inputs * math.log(false_positive_prob)) / math.pow(math.log(2), 2))
 
 
 def get_size_of_bloom_filters(sc: SparkContext, linecount_file: str, false_positive_prob: float) -> dict:
